+++ proofs/challenger/src/config.rs
@@ -1,0 +1,1 @@
+
