+++ proofs/challenger/src/sync.rs
@@ -1,0 +1,1 @@
+
