+++ proofs/primitives/Cargo.toml
@@ -11,11 +11,19 @@ repository.workspace = true
 workspace = true
 
 [dependencies]
+# workspace
+
+# alloy
 alloy-primitives.workspace = true
 alloy-contract.workspace = true
 alloy-provider.workspace = true
 alloy-sol-types.workspace = true
+
+# misc
+async-trait.workspace = true
 serde.workspace = true
+serde_json.workspace = true
+reqwest.workspace = true
 thiserror.workspace = true
 
 [features]
