+++ proofs/proposer/src/lib.rs
@@ -6,7 +6,6 @@
 mod alloy;
 mod config;
 mod error;
-mod optimism;
 mod proposer;
 mod traits;
 mod types;
@@ -14,9 +13,8 @@ mod types;
 pub use alloy::AlloyProofSystemClient;
 pub use config::ProposerConfig;
 pub use error::ProposerError;
-pub use optimism::OptimismOutputRootClient;
 pub use proposer::WorldChainProposer;
-pub use traits::{OutputRootProvider, ProofSystemClient};
+pub use traits::ProofSystemClient;
 pub use types::{ParentRef, Proposal, ProposalSubmission};
 
 #[cfg(test)]
